Add build.gradle to support Gradle Plugin Portal publishing.
A simple build.gradle that support building the plugin with the gradle
plugin helper (performing extra validation checks), and publishing to
the Gradle Plugin Portal (GPP), as well as local maven for testing.
Needs manual maintenance of the 'version' field to keep it in sync with
the master pom.xml file.  GPP publishing requires an account on the
plugins.gradle.org website.