feat: include component bom-ref
A components bom-ref has been in the CycloneDX spec since 1.1.

From the CycloneDX documentation:

"The only requirement for bom-ref is that it is unique within the BOM.
Package URL (PURL) is an ideal choice for bom-ref as it will be both
unique and readable."

Each component has a purl, so simply setting the bom-ref to purl when
the schema version is greater than 1.0.

Signed-off-by: gitphill <phill.garrett@snyk.io>